課題再々々提出

+++ src/Review01.java
@@ -2,10 +2,10 @@
 public class Review01 {
 
     public static void main(String[] args) {
-        double s = 1500;
+        int s = 1500;
         double t = 0.1;
-        double tax = tax(s,t);
-        double money = tax(s,t) + s;
+        int tax = (int)tax(s,t);
+        int money = (int)tax(s,t) + s;
         System.out.println(s + "円の商品の税込価格は" + money + "円(消費税は" + tax + "円)です。");
     }
